requirements.txt
fastapi
uvicorn
motor
python-jose
pydantic[email]
pydantic-settings
passlib==1.7.4
bcrypt==3.2.2
python-multipart
